zero-trust-fastapi-demo/services/auth_service/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
python-jose==3.3.0
pydantic==2.10.5
